refactor(cli): extract command handlers and standardize error strategy
- Extract cmdHandlers map into command-handlers.ts (index.ts: 241→93 lines)
- Extract generate helpers into generate-handlers.ts (generate.ts: 266→136 lines)
- Add CliUsageError class for exit code 2 (bad usage) errors
- Standardize plugin/target to propagate errors instead of inline try/catch
- All files now under 200-line project limit
