+++ src/cli/error-handler.ts
@@ -1,5 +1,6 @@
 import { logger } from '../utils/output/logger.js';
 import { emitJson } from './json-output.js';
+import { CliUsageError } from './cli-error.js';
 
 export interface HandleErrorOptions {
   verbose?: boolean;
@@ -8,13 +9,16 @@ export interface HandleErrorOptions {
 }
 
 /**
- * Handles CLI errors: prints message and exits with code 1.
+ * Handles CLI errors: prints message and exits.
+ * CliUsageError exits with code 2; all other errors exit with code 1.
  * In JSON mode, emits a JSON envelope to stdout instead of stderr.
  */
 export function handleError(err: Error, options?: HandleErrorOptions): never {
+  const exitCode = err instanceof CliUsageError ? 2 : 1;
+
   if (options?.json) {
     emitJson(options.command ?? 'unknown', { success: false, error: err.message });
-    return process.exit(1);
+    return process.exit(exitCode);
   }
 
   if (options?.verbose && err.stack) {
@@ -22,5 +26,5 @@ export function handleError(err: Error, options?: HandleErrorOptions): never {
   } else {
     logger.error(err.message);
   }
-  process.exit(1);
+  process.exit(exitCode);
 }
